+++ apps/macos/Sources/Flow/Views/ChannelView.swift
@@ -69,7 +69,7 @@ struct ChannelView: View {
             }
             userNames.start(db: app.db, reset: [:]) { db in
                 try Dictionary(
-                    uniqueKeysWithValues: User.fetchAll(db).map { ($0.id, $0.displayName) }
+                    uniqueKeysWithValues: User.fetchAll(db).map { ($0.id, $0.displayNameWithBadge) }
                 )
             }
             userStatuses.start(db: app.db, reset: [:]) { db in